+++ braindecode/eegneuralnet.py
@@ -14,7 +14,9 @@ import torch
 from skorch import NeuralNet
 from sklearn.metrics import get_scorer
 from skorch.callbacks import BatchScoring, EpochScoring, EpochTimer, PrintLog
+from skorch.helper import SliceDataset
 from skorch.utils import noop, to_numpy, train_loss_score, valid_loss_score, is_dataset
+
 
 from .training.scoring import (
     CroppedTimeSeriesEpochScoring,
@@ -191,7 +193,11 @@ class _EEGNeuralNet(NeuralNet, abc.ABC):
         # get kwargs from signal:
         signal_kwargs = dict()
         # Using shape to work both with torch.tensor and numpy.array:
-        if isinstance(X, mne.BaseEpochs) or (hasattr(X, "shape") and len(X.shape) >= 2):
+        if (
+            isinstance(X, mne.BaseEpochs)
+            or (hasattr(X, "shape") and len(X.shape) >= 2)
+            or isinstance(X, SliceDataset)
+        ):
             if y is None:
                 raise ValueError("y must be specified if X is array-like.")
             signal_kwargs["n_outputs"] = self._get_n_outputs(y, classes)
@@ -200,6 +206,11 @@ class _EEGNeuralNet(NeuralNet, abc.ABC):
                 signal_kwargs["n_times"] = len(X.times)
                 signal_kwargs["sfreq"] = X.info["sfreq"]
                 signal_kwargs["chs_info"] = X.info["chs"]
+            elif isinstance(X, SliceDataset):
+                self.log.info("Using SliceDataset to find signal-related parameters.")
+                Xshape = X[0].shape
+                signal_kwargs["n_times"] = Xshape[-1]
+                signal_kwargs["n_chans"] = Xshape[-2]
             else:
                 self.log.info("Using array-like to find signal-related parameters.")
                 signal_kwargs["n_times"] = X.shape[-1]